Update Flight_Carbon_Footprint_Calculator.py
This was a personal project as someone who wants to help the environment but also loves to travel. I wanted to know what I could do to offset my carbon footprint. So I looked up the calculation based on the average commercial airplane, and calculated the distance between each airport. This gives me a rough estimate on what I need to donate in order to offset that. teamtrees.org